python/decode-image-request/requirements.txt
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
opencv-python
pandas
scikit-learn